requirements.txt

aiogram==3.12.0
python-dotenv==1.0.1
pydantic==2.9.2
